Simple script for toggling screen resolution.
Some issues with streaming to a TV with a different resolution. This
works as a simple script which can be activated via Big Picture mode
to toggle the resolution on the host and make things work properly.

This was principally in place for Shadows of Mordor as its resolution
is locked to the host aspect ratio (even though it can be scaled).