Option to move (not cp/rm) games where possible.
The usual strategy for moving games between libraries is a "safe"
copy/remove one. A new 'fast' option to the move method allows us to
avoid this, which is particularly useful for very large game
installations.

A crude check to see whether the libraries exist on the same
partition/drive is employed to warn that we'll resort to an explicit
copy/remove strategy (the OS would do this anyway).