Add a materialized view to group GPU-Z records.
Grouping is present in the source data by virtue of repeated headers.
However, the current procedure strips these out. In any case, it's
more flexible to group the GPU-Z data by contiguous timestamps. This
is done in a materialized view (a view is fairly slow) which
groups records by inspecting the difference between timestamps, with
30 s being the threshold for a new recording "session".

A future update will likely introduce automatic refreshing of the
materialized view on data import, but this needs to be tested.