Enhance Fraps dataset view.
Add a timestamp for the final (Nth) frame. This is relatively slow; a
materialised version of this view has been tried and it does help but
we'd need to consider the implications of updating. It might be
overkill to use views for this, but it does avoid duplication of the
underlying data (the final time is inherit in the millisecond
attribute of the frames data).