Shift column defs inside table cpt + add AppID
Library table is specific to ...library tables... so it makes sense to
include the columns within the component.

Adding AppID (visible) temporarily but will hide it going forward.

The intent is to link pie chart slices to table rows via the AppID (as
well as a few other things for that game, e.g. store URL and header
image).