Pass data to the game modal table
Handling archived games (tbc) shows that handling this within the
functional component is not very reusable. Archived games are basically
the same(ish) but we need to access them via a different endpoint.
Creating a generic game table makes our lives easier.

Next step is to genericise the table further using react-table, except
it will be different from the library table as it's a key value table